refactor(OpenFoodFacts): move cached taxonomy logic to common_openfoodfacts
