+++ open_prices/common/openfoodfacts.py
@@ -1,4 +1,5 @@
 import datetime
+import functools
 
 import openfoodfacts
 import requests
@@ -16,6 +17,11 @@ from openfoodfacts import (
     ProductDataset,
 )
 from openfoodfacts.images import generate_image_url
+from openfoodfacts.taxonomy import (
+    create_taxonomy_mapping,
+    get_taxonomy,
+    map_to_canonical_id,
+)
 from openfoodfacts.types import JSONType
 
 OFF_CREATE_FIELDS = [
@@ -33,6 +39,51 @@ OFF_CREATE_FIELDS = [
     "unique_scans_n",
 ]
 OFF_UPDATE_FIELDS = OFF_CREATE_FIELDS + ["source", "source_last_synced"]
+
+
+# Taxonomy mapping generation takes ~200ms, so we cache it to avoid
+# recomputing it for each request.
+_cached_create_taxonomy_mapping = functools.lru_cache()(create_taxonomy_mapping)
+# Also cache the get_taxonomy function to avoid reading from disk at each
+# request.
+_cached_get_taxonomy = functools.lru_cache()(get_taxonomy)
+
+
+def normalize_taxonomized_tags(taxonomy_type: str, value_tags: list[str]) -> list[str]:
+    """Normalizes a list of tags based on the taxonomy type.
+
+    :param taxonomy_type: The type of taxonomy ('category', 'label', or
+        'origin').
+    :param value_tags: A list of tag values to normalize (e.g.,
+        ["fr: Boissons"]).
+    :raises RuntimeError: If the taxonomy type is not one of 'category',
+        'label', or 'origin'
+    :raises ValueError: If the value_tag could not be mapped to a canonical ID.
+    :return: The normalized tags (e.g., ["en:beverages"]). The order of the
+        tags is the same as the input list.
+    """
+    if taxonomy_type not in ("category", "label", "origin"):
+        raise RuntimeError(
+            f"Invalid taxonomy type: {taxonomy_type}. Expected one of 'category', 'label', or 'origin'."
+        )
+
+    # Use the cached version of the get_taxonomy function to avoid
+    # creating it multiple times.
+    category_taxonomy = _cached_get_taxonomy(taxonomy_type)
+    # the tag (category or label tag) can be provided by the mobile app in any
+    # language, with language prefix (ex: `fr: Boissons`).
+    # We need to map it to the canonical id (ex: `en:beverages`) to store it
+    # in the database.
+    # The `map_to_canonical_id` function maps the value (ex:
+    # `fr: Boissons`) to the canonical id (ex: `en:beverages`).
+    # We use the cached version of this function to avoid
+    # creating it multiple times.
+    # If the entry does not exist in the taxonomy, the tag will
+    # be set to the tag version of the value (ex: `fr:boissons`).
+    taxonomy_mapping = _cached_create_taxonomy_mapping(category_taxonomy)
+    mapped_tags = map_to_canonical_id(taxonomy_mapping, value_tags)
+    # Keep the order of the tags as they were provided
+    return [mapped_tags[k] for k in mapped_tags]
 
 
 def authenticate(username, password):
