+++ open_prices/prices/models.py
@@ -1,5 +1,4 @@
 import decimal
-import functools
 
 from django.contrib.postgres.fields import ArrayField
 from django.core.validators import MinValueValidator, ValidationError
@@ -10,18 +9,13 @@ from django.dispatch import receiver
 from django.utils import timezone
 from django_q.tasks import async_task
 from openfoodfacts.barcode import normalize_barcode
-from openfoodfacts.taxonomy import (
-    create_taxonomy_mapping,
-    get_taxonomy,
-    map_to_canonical_id,
-)
 from simple_history.models import HistoricalRecords
 
 from open_prices.challenges.models import Challenge
 
 # Import custom lookups so that they are registered
 from open_prices.common import lookups  # noqa: F401
-from open_prices.common import constants, history, utils
+from open_prices.common import constants, history, openfoodfacts, utils
 from open_prices.locations import constants as location_constants
 from open_prices.locations.models import Location
 from open_prices.prices import constants as price_constants
@@ -29,13 +23,6 @@ from open_prices.products.models import Product
 from open_prices.proofs import constants as proof_constants
 from open_prices.proofs.models import Proof
 from open_prices.users.models import User
-
-# Taxonomy mapping generation takes ~200ms, so we cache it to avoid
-# recomputing it for each request.
-_cached_create_taxonomy_mapping = functools.lru_cache()(create_taxonomy_mapping)
-# Also cache the get_taxonomy function to avoid reading from disk at each
-# request.
-_cached_get_taxonomy = functools.lru_cache()(get_taxonomy)
 
 
 class PriceQuerySet(models.QuerySet):
@@ -345,7 +332,7 @@ class Price(models.Model):
                     "Should be set to 'CATEGORY' if `category_tag` is filled",
                 )
             try:
-                self.category_tag = normalize_taxonomized_tags(
+                self.category_tag = openfoodfacts.normalize_taxonomized_tags(
                     "category", [self.category_tag]
                 )[0]
             except ValueError as e:
@@ -364,7 +351,7 @@ class Price(models.Model):
                     )
                 else:
                     try:
-                        self.labels_tags = normalize_taxonomized_tags(
+                        self.labels_tags = openfoodfacts.normalize_taxonomized_tags(
                             "label", self.labels_tags
                         )
                     except ValueError as e:
@@ -382,7 +369,7 @@ class Price(models.Model):
                     )
                 else:
                     try:
-                        self.origins_tags = normalize_taxonomized_tags(
+                        self.origins_tags = openfoodfacts.normalize_taxonomized_tags(
                             "origin", self.origins_tags
                         )
                     except ValueError as e:
@@ -735,40 +722,3 @@ def price_post_delete_decrement_counts(sender, instance, **kwargs):
         Location.objects.filter(id=instance.location.id).update(
             price_count=F("price_count") - 1
         )
-
-
-def normalize_taxonomized_tags(taxonomy_type: str, value_tags: list[str]) -> list[str]:
-    """Normalizes a list of tags based on the taxonomy type.
-
-    :param taxonomy_type: The type of taxonomy ('category', 'label', or
-        'origin').
-    :param value_tags: A list of tag values to normalize (e.g.,
-        ["fr: Boissons"]).
-    :raises RuntimeError: If the taxonomy type is not one of 'category',
-        'label', or 'origin'
-    :raises ValueError: If the value_tag could not be mapped to a canonical ID.
-    :return: The normalized tags (e.g., ["en:beverages"]). The order of the
-        tags is the same as the input list.
-    """
-    if taxonomy_type not in ("category", "label", "origin"):
-        raise RuntimeError(
-            f"Invalid taxonomy type: {taxonomy_type}. Expected one of 'category', 'label', or 'origin'."
-        )
-
-    # Use the cached version of the get_taxonomy function to avoid
-    # creating it multiple times.
-    category_taxonomy = _cached_get_taxonomy(taxonomy_type)
-    # the tag (category or label tag) can be provided by the mobile app in any
-    # language, with language prefix (ex: `fr: Boissons`).
-    # We need to map it to the canonical id (ex: `en:beverages`) to store it
-    # in the database.
-    # The `map_to_canonical_id` function maps the value (ex:
-    # `fr: Boissons`) to the canonical id (ex: `en:beverages`).
-    # We use the cached version of this function to avoid
-    # creating it multiple times.
-    # If the entry does not exist in the taxonomy, the tag will
-    # be set to the tag version of the value (ex: `fr:boissons`).
-    taxonomy_mapping = _cached_create_taxonomy_mapping(category_taxonomy)
-    mapped_tags = map_to_canonical_id(taxonomy_mapping, value_tags)
-    # Keep the order of the tags as they were provided
-    return [mapped_tags[k] for k in mapped_tags]
